+++ mod/TTT.Roles/Helper/ChatManager.cs
