+++ mod/TTT.Roles/Helper/InfoManager.cs
